pyproject.toml
[tool.poetry]
name = "architxt"
version = "0.1.0"
description = "ArchiTXT is a tool for structuring textual data into a valid database model. It is guided by a meta-grammar and uses an iterative process of tree rewriting."
authors = [
    "Nicolas Hiot <nicolas.hiot@univ-orleans.fr>",
    "Mirian Halfeld-Ferrari <mirian@univ-orleans.fr>",
    "Jacques Chabin <jchabin@univ-orleans.fr>"
]
readme = "README.md"

[tool.poetry.scripts]
architxt = "architxt.cli:main"

[tool.poetry-dynamic-versioning]
enable = true

[build-system]
requires = [
    "poetry-core>=1.0.0",
    "poetry-dynamic-versioning>=1.0.0,<2.0.0",
]
build-backend = "poetry_dynamic_versioning.backend"

[tool.poetry.dependencies]
python = ">=3.10,<3.13"

# Core dependencies
tqdm = ">=4.60"
nltk = "*"
chatette = ">=1.6.2"
unidecode = "*"
spacy-alignments = "^0.9.1"
requests = "*"
numpy = "^2.1.3"
levenshtein = "*"
more-itertools = "*"
bratlib = { git = "https://github.com/swfarnsworth/bratlib.git" }
mlflow = "^2.6.0"
psutil = "^6.1.0"
typer = "^0.15.1"
anyio = "^4.6.0"
rich = "^13.9.4"
antlr4-python3-runtime = ">=4.13,<5"

# Optional UI dependencies
streamlit = { version = ">=1.39.0", optional = true }
streamlit-agraph = { version = "*", optional = true }
streamlit-tags = { version = "*", optional = true }

[tool.poetry.extras]
ui = ["streamlit", "streamlit-agraph", "streamlit-tags"]

[tool.poetry.group.dev.dependencies]
wheel = "^0.45.0"
devtools = "^0.12.0"
py-spy = "~0.4.0"
pre-commit = "^4.0.1"
datasets = "*"

pylint = "^3.3"
mypy = ">=0.910,<1.14"
ruff = "^0.8.0"

# Typing stubs
pandas-stubs = "*"
types-tqdm = "*"
types-requests = "*"

# Tools for generating grammar parsers
antlr4-tools = "^0.2.1"

[tool.poetry.group.test.dependencies]
pytest = "^8.3"
pytest-cov = "*"
pytest-xdist = "*"
pytest-mock = "*"

beartype = "^0.19.0"
pytest-beartype = "^0.2.0"

hypothesis = "^6.68.0"

# Async testing libraries
asyncio = "^3.4.0"
trio = "^0.27.0"

[tool.pytest.ini_options]
minversion = "7.0"
junit_family = "legacy"
testpaths = ["architxt", "tests"]
addopts = [
    "--doctest-modules",
    "--beartype-skip-packages=architxt/grammar"
]

[tool.coverage.run]
source = ["architxt"]

[tool.coverage.report]
exclude_also = [
    "if TYPE_CHECKING:"
]

[tool.ruff]
target-version = "py310"
line-length = 120
src = ["architxt", "tests"]
exclude = ["architxt/grammar"]

[tool.ruff.lint]
select = [
    "E4", # Pycodestyle error codes that are safe to auto-fix
    "E7",
    "E9",
    "C4", # Comprehensions
    "F", # Flake8 rules
    "I", # isort rules for import ordering
    "ICN", # Import conventions
    "TID", # Tidy imports
    "UP", # pyupgrade (for safe code modernization)
    "C4", # Comprehension simplifications
    "SIM", # Code simplifications
    "LOG", # Logger usage
    "PIE", # Miscellaneous
    "RUF", # Ruff-specific auto-fixable rules
    "PD", # Pandas checks
    "NPY", # NumPy checks
    "PERF", # Performance anti-pattern checks
    "FURB", # Old Python code
    "PT", # Pytesst checks
    "RET", # Returns
    "PTH", # Prefer Pathlib
]

[tool.ruff.format]
quote-style = "preserve"
